web/requirements.txt
asgiref==3.7.2
black==24.1.1
click==8.1.7
channels==3.0.4
channels-redis==3.3.1
Django==4.2.9
djangorestframework==3.14.0
exceptiongroup==1.2.0
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytest==8.0.0
pytest-django==4.8.0
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
requests==2.31.0
channels_redis

